Allow setting `dicomPS315EOptions` to 'Off' in mapping specification. This is for running dcm-organize without de-identification, e.g. for counting validation issues, or mapping tags.
